Implementado o gerenciamento de funcionários. Cadastra, Edita e Inativa Implementado um conversor de CPF

+++ trunk/primoslavacar/src/main/java/com/primos/lavacar/model/FuncionarioModel.java
@@ -62,7 +62,7 @@ public interface FuncionarioModel {
 	 * @throws SQLException
 	 * @throws ClassNotFoundException
 	 */
-	public void inativar(Integer matricula) throws SQLException,
+	public void inativar(Long  matricula) throws SQLException,
 			ClassNotFoundException;
 
 	/**
@@ -73,7 +73,7 @@ public interface FuncionarioModel {
 	 * @throws SQLException
 	 * @throws ClassNotFoundException
 	 */
-	public void alterar(FuncionarioViewBean funcionario, Integer matriculaAntiga)
+	public void alterar(FuncionarioViewBean funcionario, Long matriculaAntiga)
 			throws SQLException, ClassNotFoundException;
 
 	/**
@@ -126,7 +126,7 @@ public interface FuncionarioModel {
 	 * @throws SQLException
 	 * @throws ClassNotFoundException
 	 */
-	public void removerColaborador(Integer matricula) throws SQLException,
+	public void removerColaborador(Long matricula) throws SQLException,
 			ClassNotFoundException;
 
 	/**
